Co-locate API endpoints and query keys in feature modules
Move endpoint definitions and query key factories from the central
lib/api.ts and lib/queryClient.ts files into co-located api.ts files
within 6 feature modules: artists, venues, shows, releases, labels,
and festivals.

This reduces the friction of adding new endpoints from 5-file edits
to 2-file edits (feature api.ts + feature hooks). Central files
re-export from feature modules for full backward compatibility.

- Extract API_BASE_URL to lib/api-base.ts to break circular imports
- Create features/{artists,venues,shows,releases,labels,festivals}/api.ts
  with co-located endpoints and query keys
- Update 30 hook files to import from co-located ../api
- Update lib/api.ts to re-export feature endpoints (backward compat)
- Update lib/queryClient.ts to re-export feature query keys
- Update 18 test files to mock the new feature api modules
- 72 endpoints and 40+ query keys moved across 6 feature modules
- All 2408 tests pass, build succeeds with no errors

Closes PSY-192
