+++ frontend/features/artists/hooks/useArtists.test.tsx
@@ -9,26 +9,22 @@ const mockApiRequest = vi.fn()
 // Mock the api module
 vi.mock('@/lib/api', () => ({
   apiRequest: (...args: unknown[]) => mockApiRequest(...args),
-  API_ENDPOINTS: {
-    ARTISTS: {
-      LIST: '/artists',
-      CITIES: '/artists/cities',
-      GET: (artistId: string | number) => `/artists/${artistId}`,
-      SHOWS: (artistId: string | number) => `/artists/${artistId}/shows`,
-    },
-  },
   API_BASE_URL: 'http://localhost:8080',
 }))
 
-// Mock queryClient module
-vi.mock('@/lib/queryClient', () => ({
-  queryKeys: {
-    artists: {
-      list: (filters?: Record<string, unknown>) => ['artists', 'list', filters],
-      cities: ['artists', 'cities'],
-      detail: (id: string | number) => ['artists', 'detail', String(id)],
-      shows: (artistId: string | number) => ['artists', 'shows', String(artistId)],
-    },
+// Mock the feature api module
+vi.mock('../api', () => ({
+  artistEndpoints: {
+    LIST: '/artists',
+    CITIES: '/artists/cities',
+    GET: (artistId: string | number) => `/artists/${artistId}`,
+    SHOWS: (artistId: string | number) => `/artists/${artistId}/shows`,
+  },
+  artistQueryKeys: {
+    list: (filters?: Record<string, unknown>) => ['artists', 'list', filters],
+    cities: ['artists', 'cities'],
+    detail: (id: string | number) => ['artists', 'detail', String(id)],
+    shows: (artistId: string | number) => ['artists', 'shows', String(artistId)],
   },
 }))
