+++ frontend/features/artists/hooks/useArtists.ts
@@ -7,8 +7,8 @@
  */
 
 import { useQuery, keepPreviousData } from '@tanstack/react-query'
-import { apiRequest, API_ENDPOINTS } from '@/lib/api'
-import { queryKeys } from '@/lib/queryClient'
+import { apiRequest } from '@/lib/api'
+import { artistEndpoints, artistQueryKeys } from '../api'
 import type { CityState } from '@/components/filters'
 import type {
   Artist,
@@ -36,11 +36,11 @@ export function useArtists(options: UseArtistsOptions = {}) {
 
   const queryString = params.toString()
   const endpoint = queryString
-    ? `${API_ENDPOINTS.ARTISTS.LIST}?${queryString}`
-    : API_ENDPOINTS.ARTISTS.LIST
+    ? `${artistEndpoints.LIST}?${queryString}`
+    : artistEndpoints.LIST
 
   return useQuery({
-    queryKey: queryKeys.artists.list(cities ? { cities } : undefined),
+    queryKey: artistQueryKeys.list(cities ? { cities } : undefined),
     queryFn: async (): Promise<ArtistsListResponse> => {
       return apiRequest<ArtistsListResponse>(endpoint, {
         method: 'GET',
@@ -56,9 +56,9 @@ export function useArtists(options: UseArtistsOptions = {}) {
  */
 export function useArtistCities() {
   return useQuery({
-    queryKey: queryKeys.artists.cities,
+    queryKey: artistQueryKeys.cities,
     queryFn: async (): Promise<ArtistCitiesResponse> => {
-      return apiRequest<ArtistCitiesResponse>(API_ENDPOINTS.ARTISTS.CITIES, {
+      return apiRequest<ArtistCitiesResponse>(artistEndpoints.CITIES, {
         method: 'GET',
       })
     },
@@ -79,9 +79,9 @@ export function useArtist(options: UseArtistOptions) {
   const { artistId, enabled = true } = options
 
   return useQuery({
-    queryKey: queryKeys.artists.detail(artistId),
+    queryKey: artistQueryKeys.detail(artistId),
     queryFn: async (): Promise<Artist> => {
-      return apiRequest<Artist>(API_ENDPOINTS.ARTISTS.GET(artistId), {
+      return apiRequest<Artist>(artistEndpoints.GET(artistId), {
         method: 'GET',
       })
     },
@@ -119,11 +119,11 @@ export function useArtistShows(options: UseArtistShowsOptions) {
 
   const queryString = params.toString()
   const endpoint = queryString
-    ? `${API_ENDPOINTS.ARTISTS.SHOWS(artistId)}?${queryString}`
-    : API_ENDPOINTS.ARTISTS.SHOWS(artistId)
+    ? `${artistEndpoints.SHOWS(artistId)}?${queryString}`
+    : artistEndpoints.SHOWS(artistId)
 
   return useQuery({
-    queryKey: [...queryKeys.artists.shows(artistId), timeFilter],
+    queryKey: [...artistQueryKeys.shows(artistId), timeFilter],
     queryFn: async (): Promise<ArtistShowsResponse> => {
       return apiRequest<ArtistShowsResponse>(endpoint, {
         method: 'GET',
