+++ frontend/features/shows/api.ts
@@ -1,0 +1,54 @@
+/**
+ * Shows API Configuration
+ *
+ * Co-located endpoint definitions and query keys for the shows feature.
+ * Imported by show hooks and re-exported from lib/api.ts and lib/queryClient.ts
+ * for backward compatibility.
+ */
+
+import { API_BASE_URL } from '@/lib/api-base'
+
+// ============================================================================
+// Endpoints
+// ============================================================================
+
+export const showEndpoints = {
+  SUBMIT: `${API_BASE_URL}/shows`,
+  UPCOMING: `${API_BASE_URL}/shows/upcoming`,
+  CITIES: `${API_BASE_URL}/shows/cities`,
+  GET: (showId: string | number) => `${API_BASE_URL}/shows/${showId}`,
+  UPDATE: (showId: string | number) => `${API_BASE_URL}/shows/${showId}`,
+  DELETE: (showId: string | number) => `${API_BASE_URL}/shows/${showId}`,
+  UNPUBLISH: (showId: string | number) =>
+    `${API_BASE_URL}/shows/${showId}/unpublish`,
+  MAKE_PRIVATE: (showId: string | number) =>
+    `${API_BASE_URL}/shows/${showId}/make-private`,
+  PUBLISH: (showId: string | number) =>
+    `${API_BASE_URL}/shows/${showId}/publish`,
+  SET_SOLD_OUT: (showId: string | number) =>
+    `${API_BASE_URL}/shows/${showId}/sold-out`,
+  SET_CANCELLED: (showId: string | number) =>
+    `${API_BASE_URL}/shows/${showId}/cancelled`,
+  MY_SUBMISSIONS: `${API_BASE_URL}/shows/my-submissions`,
+  // Export endpoint (dev only)
+  EXPORT: (showId: string | number) =>
+    `${API_BASE_URL}/shows/${showId}/export`,
+  // Show report endpoints
+  REPORT: (showId: string | number) =>
+    `${API_BASE_URL}/shows/${showId}/report`,
+  MY_REPORT: (showId: string | number) =>
+    `${API_BASE_URL}/shows/${showId}/my-report`,
+} as const
+
+// ============================================================================
+// Query Keys
+// ============================================================================
+
+export const showQueryKeys = {
+  all: ['shows'] as const,
+  list: (filters?: Record<string, unknown>) =>
+    ['shows', 'list', filters] as const,
+  cities: (timezone?: string) => ['shows', 'cities', timezone] as const,
+  detail: (id: string) => ['shows', 'detail', id] as const,
+  userShows: (userId: string) => ['shows', 'user', userId] as const,
+} as const
