+++ frontend/features/shows/hooks/useShowReports.test.tsx
@@ -7,13 +7,14 @@ const mockInvalidateShowReports = vi.fn()
 
 vi.mock('@/lib/api', () => ({
   apiRequest: (...args: unknown[]) => mockApiRequest(...args),
-  API_ENDPOINTS: {
-    SHOWS: {
-      REPORT: (showId: string | number) => `/shows/${showId}/report`,
-      MY_REPORT: (showId: string | number) => `/shows/${showId}/my-report`,
-    },
-  },
   API_BASE_URL: 'http://localhost:8080',
+}))
+
+vi.mock('../api', () => ({
+  showEndpoints: {
+    REPORT: (showId: string | number) => `/shows/${showId}/report`,
+    MY_REPORT: (showId: string | number) => `/shows/${showId}/my-report`,
+  },
 }))
 
 vi.mock('@/lib/queryClient', () => ({
