+++ frontend/features/shows/hooks/useShowSubmit.test.tsx
@@ -11,10 +11,11 @@ const mockInvalidateSavedShows = vi.fn()
 // Mock the api module
 vi.mock('@/lib/api', () => ({
   apiRequest: (...args: unknown[]) => mockApiRequest(...args),
-  API_ENDPOINTS: {
-    SHOWS: {
-      SUBMIT: '/shows',
-    },
+}))
+
+vi.mock('../api', () => ({
+  showEndpoints: {
+    SUBMIT: '/shows',
   },
 }))
